Add ProjectStream to depository
sales_data_sample.csv is required for ProjectStream to run. It is a predictive model that clusters groups of customers for a client and determines to what confidence we can predict their spending trends. This model was able to do this between 95-97% confidence, and was successfully able to cluster the client's customers into 5 clusters based on the most important variables in predicting their spending. These variables include region, product, product cost, product quantity, and territory.